refactor(frontend-ui): Upgrade form controls for signal forms
