+++ frontend/ethereal-ui/src/lib/components/primitives/index.ts
@@ -1,10 +1,5 @@
-export * from './checkbox-input.component';
 export * from './currency-amount.component';
 export * from './glass-panel.component';
 export * from './icon-button.component';
-export * from './pill-toggle.component';
 export * from './quality-badge.component';
-export * from './search-input.component';
-export * from './select-input.component';
 export * from './symbol-icon.component';
-export * from './text-input.component';
