+++ frontend/ethereal-ui/src/public-api.ts
@@ -1,3 +1,4 @@
+export * from './lib/components/forms';
 export * from './lib/components/market';
 export * from './lib/components/primitives';
 export * from './lib/components/shell';
